backend/requirements.txt
python-dotenv
flask
flask-cors
pillow
google-generativeai
google-cloud-storage